+++ docs/changeLog.md
@@ -2,9 +2,10 @@
 Current version: 1.1.0-beta
 
 ## 1.1.0-beta
+- Added copyable entities
 - Added the texture manager and shared textures
 - Reverted the change where the input element would round the value
-- Added support for dragging entities (TBD if draggable is true)
+- Added support for dragging entities if the draggable attribute is true
 - Added the hoverForBanks attribute that determins whether the mouse cursor has to hover over the bank inorder to change the value
 
 ## 1.0.1-beta